pyproject.toml
[tool.poetry]
name = "nudging"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.9,<3.10"
pandas = "^1.2.4"
numpy = "^1.20.2"
openpyxl = "^3.0.7"
pyreadstat = "^1.1.0"
scipy = "^1.6.3"
scikit-learn = "^0.24.2"
seaborn = "^0.11.1"
statsmodels = "^0.12.2"
CausalInference = "^0.1.3"
h5py = "^3.4.0"
PyYAML = "^5.4.1"
tqdm = "^4.62.3"

[tool.poetry.dev-dependencies]
pylint = "^2.8.2"
flake8 = "^3.9.2"
pytest = "^6.2.5"
pytest-cov = "^2.12.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
